Refactor HomeController and home.blade.php to retrieve races from a testing API
This commit refactors the HomeController and home.blade.php files to retrieve races from an API instead of directly querying the database. The code that retrieves races from the database has been commented out, and new code has been added to make an HTTP GET request to the API and retrieve the races. The view has also been updated to display the race details correctly.